bump r2dbc to 1.0.0

+++ the.orm/src/main/kotlin/io/the/orm/dbio/r2dbc/R2dbcResult.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.reactive.awaitSingle
 import org.reactivestreams.Publisher
 
 class R2dbcResult(private val result: io.r2dbc.spi.Result) : DBResult {
-    override suspend fun rowsUpdated(): Int = result.rowsUpdated.awaitSingle()
+    override suspend fun rowsUpdated(): Long = result.rowsUpdated.awaitSingle()
 
     override suspend fun <T : Any> map(mappingFunction: (t: DBRow) -> T): Flow<T> {
         val mapped: Publisher<T> = try {
