+++ the.orm/src/main/kotlin/io/the/orm/exp/testing/MockConnectionProvider.kt
@@ -43,8 +43,8 @@ class MockStatement(val sql: String) : Statement {
     }
 }
 
-data class MockDBResult(val rows: List<DBRow> = listOf(), val rowsUpdated: Int = 0, val id: Long = 0) : DBResult {
-    override suspend fun rowsUpdated(): Int = rowsUpdated
+data class MockDBResult(val rows: List<DBRow> = listOf(), val rowsUpdated: Long = 0, val id: Long = 0) : DBResult {
+    override suspend fun rowsUpdated(): Long = rowsUpdated
     override suspend fun getId(): Long = id
 
     override suspend fun <T : Any> map(mappingFunction: (t: DBRow) -> T): Flow<T> {
